Extract configuration file reading from CatalogMergeTool for reuse
